update: refactored recursive residual groups + corresponding unit tests
